[CBR-267] Split unit tests into submodules
